// src/interfaces/api.ts
export type DatabaseQueryResponseType = {
  data?: any;
  error?: any;
};